refactor(react): rename src/pages/upload component
